move RecordIndexUpdater to its own file
Clean things up, plus avoid accidentally adding it as a namespace-nested class in whatever includes Kithe::Indexable
